requirements.txt
opencv-python
deepface
mysql-connector-python
pytest